+++ test/InterpreterSpec.hs
@@ -2,7 +2,7 @@ module InterpreterSpec (main, spec) where
 
 import Test.Hspec
 import Interpreter.Data.Environment
-import Interpreter.System.BinaryOperation
+import Interpreter.System.Operation
 import Interpreter.System.Types
 import Interpreter.System.Evaluator
 import Utils.Data.Result
